requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
dataclasses-json==0.6.7
deepgram-sdk==4.0.0
deprecation==2.1.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-whisper==20240930
packaging==25.0
playsound==1.3.0
propcache==0.3.1
PyAudio==0.2.11
pycparser==2.22
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
sniffio==1.3.1
sounddevice==0.5.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
triton==3.3.0
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
webrtcvad==2.0.10
websockets==15.0.1
yarl==1.20.0
